Update Dockerfile for midpoint data ingestion
Removed `boot.sh` since it applies for the final application.
For the midproject checkpoint, (1) data will be downloaded
from source and uploaded to S3, and (2) a MySQL database will
be created either locally or in AWS RDS.

`docker run` will be used for each command, so `boot.sh` is
not needed.